Refactor middleware imports: update admin controller and routes to use admin middleware for token handling
